Fixing SDNVPN tempest API tests
The code to run Tempest tests has changed in SDNVPN between Euphrates
and Fraser, causing these tests to fail. Moreover, due to a bug in
SDNVPN, not all bgpvpn API tests are currently being run. This is being
addressed by https://gerrit.opnfv.org/gerrit/#/c/57299/ and needs to be
cherrypicked to Fraser 6.1 for inclusion in OVP 2.0.

Change-Id: I0a7575e8a9ebc23af5edaa26c1a9ce1e8149919c
Signed-off-by: Georg Kunz <georg.kunz@ericsson.com>