Update patch set 1
Patch Set 1:

I just merged the user guide update. I will push an update to this patchset to add the description to the user guide.

I was not aware of the problem regarding the external network because, according to my understanding, an external network is actually required to create and assign floating IPs. If that's not the case, we need to address this by means of documentation, e.g., describing that the network which provides external connectivity needs to be created with that flag as part of the system prepartion. If I am misunderstanding something here, please let me know.

Patch-set: 1
Label: Verified=0