Update patch set 2
Patch Set 2:

We have previously discussed whether or not Dovetail should select individual test cases from the overall set of tests available in Functest and it is important to return to this discussion.

I see a difference between the purpose of a functional test framework like Functest and a compliance program: Functest tests that a system works and behaves correctly - given the functionality and configuration of the system. This means, if the system is not configured with a block storage backend, then it makes no sense - from the perspective of functional testing - to run the block storage tests and report FAIL. In this context, it makes sense to skip the tests altogether. However, all functionality which is enabled should be tested and tests should pass.

A compliance program on the other hand defines WHICH functionality must be present AND that this functionality muss be correct. So, if it is decided in OVP - for example - that block storage is part of the system capabilities, than it must be present, tested, and pass the tests. We cannot skip those tests and report PASS. From this perspective, a compliance program must explicitly define which tests are expected to be executed and pass. The fact that Tempest by design can skip test cases is due to the fact that it is a functional testing framework and not a compliance test framework. RefStack for instance reads a list of test cases according to an Interop "guideline".

As already touched upon some time ago, Dovetail could either also give a list of individual test cases to Functest for execution (as done with the tempest_custom tests) or call a larger test suite in Functest (such as patrole or the trunk port tests) and compare the results against a list of tests we need to see pass (and which tests are skippable etc.). Very likely, we need to use both mechanisms.

Patch-set: 2
Label: Verified=0