Update patch set 2
Patch Set 2:

And there is also an interesting point. OPNFV only defines the OpenStack version as requirement. And here Dovetail asks for a full complete sets of services (e.g. neutron features).

It means that OPNFV installers may not pass the certification if the test list is not too short.

Patch-set: 2