Create specific flavor for neutron-l3-agent HA test
The neutron-l3-agent test assumes that a default image and flavor are
available. On current OPNFV deployments, these resources do not exist
anymore by default and hence the test case fails.

This patch enables Dovetail to create a specific flavor as part of the
test case config. This is temporary workaround until the test case
itself takes care of checking and creating all resources needed. A
ticket for the HA project has been created accordingly:
https://jira.opnfv.org/browse/HA-36

Change-Id: I6523d8411fe853a5645566b2fabe1409b5715491
Signed-off-by: Georg Kunz <georg.kunz@ericsson.com>