Update patch set 1
Patch Set 1:

Your feedback is appreciated and valuable. I think we agree in this case as we both want to remove a test which is not gated.

Ideally, Dovetail just calls a test case defined in Functest (see the trunk port tests, vIMS, etc.). In other cases (Tempest only), we pass a list of tests to Functest. Isn't that similar to what RefStack is doing? Selecting a list of tests and asking Tempest to run them?

Patch-set: 1
Reviewer: Gerrit User 585 <585@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0