Update patch set 1
Patch Set 1:

> @Xudan: both, the API and the scenario tests pass fine on my
 > virtual Apex deployment. Can you please check the tests on your
 > compass deployment? I really have the impression the CI results are
 > so shaky because the dovetail suite runs after all other tests have
 > already loaded the system for hours.

Sorry for gave the wrong info link about the ipv6 on yesterday's meeting. The right info is given in the release note -- known issue https://wiki.opnfv.org/display/dovetail/Running+history+for+the+dovetail+tool#Runninghistoryforthedovetailtool-4.KnownIssuesList
I don't think these test cases failed because of the SUT becomes shaky after running a lot of test cases. I have checked compass and Fuel odl scenarios, and the reason for the failed ipv6 test cases are the same which is "AssertionError: False is not true : Timed out waiting for 2003::f816:3eff:fec0:3f3d to become reachable from 100.64.200.114". And we can confirm this with the ODL guys to find out that if they still can't support that.

Patch-set: 1