Update patch set 1
Patch Set 1:

> Hi Panos,
 > it seems the Functest code has been refactored in the meantime, but
 > the same basic principle remains: Compare [1] and [2].
 > 
 > The current mechanism uses a patch.yaml [3] file to set flavor
 > extra specs based on a substring of the scenario name. This means
 > for your task that you just need to make sure that Dovetail exports
 > a DEPLOY_SCENARIO containing "ovs" if your enable_dpdk flag is
 > enabled.
 > 
 > Actually, please rename the Dovetail command line option to
 > enable_hugepages which is more generic.
 > 
 > [1] https://github.com/opnfv/functest/blob/stable/euphrates/functest/opnfv_tests/openstack/vping/vping_base.py#L159
 > [2]https://github.com/opnfv/functest/blob/stable/fraser/functest/opnfv_tests/openstack/vping/vping_base.py#L125
 > [3] https://github.com/opnfv/functest/blob/stable/fraser/functest/ci/config_patch.yaml


Hi Georg,

I see your point and yes it seems that there is a refactoring from Functest guys.
But, I think so that this refactoring doesn't affect our side.

I can change the option name from '--enable-dpdk- to '--enable-hugepage', but I think that the whole ticket should change description/title, because as you said the hugepage enabling is something more generic and this is not related to OVS and DPDK. Moreover, my concern from functional point of view is that I will put the 'ovs' as Deploy_scenario in order for the flavor to be updated with hugepage, even if not OVS is really valid. Of course, I know that the users who intent to use this improvement will be only the testers and this is not for 'commercial' use.

Therefore, the final decision is yours :) What do you prefer?

Patch-set: 1