Update patch set 1
Patch Set 1:

You are right that mapping a new CLI option such as "enable-hugepages" to a specific scenario name is not a clean solution. I wanted to avoid that users of Dovetail need to know about scenarios, but scenarios are t he primary configuration mechanism in OPNFV, so we should not try to hide this with shaky hacks.

Instead, a cleaner way would be the following: provide a command line option for Dovetail to specify a scenario name which Dovetail then exports to the test tools. Currrent, the scenario name is hardcoded to "unknown" which is also not clean [1]. We then need to add a section in the user guide explaining the implications of using a given scenario.

[1] https://github.com/opnfv/dovetail/blob/master/etc/conf/functest_config.yml#L14

Patch-set: 1