Test case descriptions for Neutron Trunk Port tests
Adding a high-level description of the Neutron trunk port tempest tests.
As agreed on the last Dovetail call, we are aiming for a relatively low
level of detail and instead point to available documentation. As the
tempest tests themselves are quite simple of read, this is the best,
most concise, and up-to-date version of documentation we can find.

JIRA: DOVETAIL-684

Change-Id: I76801b34f974bf7c77fded3ffa5588b7e5b2803c
Signed-off-by: Georg Kunz <georg.kunz@ericsson.com>