Update patch set 2
Patch Set 2: Code-Review-1

I'm not involved in the discussions / decisions of this ticket.
However, during testing of this patch , I got the following outcome:

bash-4.4# dovetail list
- debug
    optional
- onap.1.0.0
    optional
- ovp.next
    mandatory
    optional
- proposed_tests
    mandatory
    optional
- healthcheck
    optional

My opinion is that it could be better if the list of the test cases are appeared as before.
Do you agree?

Patch-set: 2
Label: Code-Review=-1