Update patch set 1
Patch Set 1:

I'd like to discuss this in the broader team. This change has two implications:
i) How many release of OpenStack do we want to cover in terms of backwards compatibility? API 2.5 was introduced in Ocata, i.e. one release before Pike which is the basis for Fraser. 

ii) I cannot recall if these two tests failed or were skipped. In case  of failure, Functest marks the overall test as failed as well - in contrast to skipped tests. In this case, our webportal needs to determine that an overall testcase still passed even if a few of its subtest failed.

Patch-set: 1