Update patch set 1
Patch Set 1:

Georg. We don't have any installer dependencies in master.
The key issue about scenario is the global OPNFV design. We are cleaning the staff which should have been in Releng instead of Functest (see CI_LOOP as the easiest example).

Here Dovetail runs the selected testcase which is better than the design we are cleaning. More when several features such as Doctor are not defined by specific scenarios.

Patch-set: 1