Test case descriptions for SNAPS smoke tests
Adding a high-level description of the SNAPS smoke tests.
As agreed on the last Dovetail call, we are aiming for a relatively low
level of detail and instead point to available documentation. As the
tests themselves are quite simple of read, this is the best,
most concise, and up-to-date version of documentation we can find.

JIRA: DOVETAIL-630

Change-Id: I38b072e29ec4cfe50baf0ac3350ea1a1d3e8d67a
Signed-off-by: dimitris.tsiolakis <dmts@intracom-telecom.com>