Remove 2 sub test cases from neutron_trunk check list
1. neutron.tests.tempest.api.test_trunk.TrunkTestJSON.test_create_update_trunk
This test case will get the value of key 'revision_number' [1],
While this is a new extension for Newton and not support by Mitaka.
Please refer to the history of neutron github and the differences between
mitaka tag [3] (which has no file revisions.py) and newton [4].

2. neutron.tests.tempest.api.test_trunk.TrunkTestJSON.test_show_trunk_has_project_id
This test case requires extension 'project-id' [5], while this is also a new
extension for Newton. Please refer to the history of neutron [6] and the difference
between mitaka tag [3] (which has no file project_id.py) and newton [7].

Also it is a little unreasonable to check both tenant_id and project_id.
For some products, they may only have either tenant_id or project_id (not both).
Has sent an email to the neutron_trunk team but get no feedback until now.

[1] https://github.com/openstack/neutron/blob/stable%2Fpike/neutron/tests/tempest/api/test_trunk.py#L123
[2] https://github.com/openstack/neutron/commits/newton-eol/neutron/extensions/revisions.py
[3] https://github.com/openstack/neutron/tree/mitaka-eol/neutron/extensions
[4] https://github.com/openstack/neutron/blob/newton-eol/neutron/extensions/revisions.py
[5] https://github.com/openstack/neutron/blob/stable%2Fpike/neutron/tests/tempest/api/test_trunk.py#L110
[6] https://github.com/openstack/neutron/commits/newton-eol/neutron/extensions/project_id.py
[7] https://github.com/openstack/neutron/blob/master/neutron/extensions/project_id.py

JIRA: DOVETAIL-723

Change-Id: Idb4a56d47148b14ae4674549f98b1be65e9b8f15
Signed-off-by: xudan <xudan16@huawei.com>