Test case descriptions for SNAPS smoke tests
Adding a high-level description of the SNAPS smoke tests.
As agreed on the last Dovetail call, we are aiming for a relatively low
level of detail and instead point to available documentation. As the
tests themselves are quite simple of read, this is the best,
most concise, and up-to-date version of documentation we can find.

JIRA: DOVETAIL-630

Change-Id: I985592c55fd415e254e70db6c6480c7bf3670a29
Signed-off-by: dimitris.tsiolakis <dmts@intracom-telecom.com>