Update patch set 1
Patch Set 1:

The respective references for 'identity v2' should be removed from documentation rst files as well, right?
(Alternatively, a note should be added, indicating that the version 2 is not supported from functest anymore)
What do you believe? Which altenative is more efficient?
I will do this as part of the next update of this patchset.

Patch-set: 1
Label: Verified=0