[WIP]Add User Guide per installer - APEX
Trying to install the dovetail project on APEX, some instructions
have been found that are not clear enough.
The aim of introduction of new folder and user guide documents for
each installer is to make a dovetail deployment easier, using tips
and advices that are applicable for each installer respectively.

JIRA: DOVETAIL-743

Change-Id: I907b694d859164031b170cb0c7bb4f69b9ec889b
Signed-off-by: Panagiotis Karalis <pkaralis@intracom-telecom.com>