Remove neutron trunk test cases with href links
There is a bug with neutron when using href links to get url.
This is fixed with Pike.

This is the patch,
https://review.openstack.org/#/c/356707/

These are the related files of Ocata and Pike,
https://github.com/openstack/neutron/blob/stable/ocata/neutron/api/api_common.py#L68
https://github.com/openstack/neutron/blob/stable/pike/neutron/api/api_common.py#L95

There are 2 test cases in neutron trunk related to this.
Remove these 2 test cases since Dovetail should support 4 versions from M to P.

Change-Id: I8aac88e4688157def1d6642eaa64bba56dde49ba
Signed-off-by: xudan <xudan16@huawei.com>