New PTL for Dovetail
This patch is used as a basis for voting on the new Dovetail PTL.  After
the nomination period [1] has ended, there is one candidate [2].

In order to complete the PTL election process, all Dovetail committers
shall review this patch and cast a vote (-1, +1). The voting period ends
in a week from now on October 24. The vote passes if a majority of
committers +1 this patch.

[1] https://lists.opnfv.org/g/opnfv-tech-discuss/message/22133
[2] https://lists.opnfv.org/g/opnfv-tech-discuss/message/22145

Change-Id: Iacf53f5c0bbe7e0c5f3c47131569be72ed8623ec
Signed-off-by: Georg Kunz <georg.kunz@ericsson.com>