Update patch set 4
Patch Set 4:

> (1 comment)
 > 
 > Lots of unclear blacklists and the git commit message doesn't
 > mention them. Could you please explain why?
 > 
 > Dovetail should conform with Functest and mustn't decide how to
 > override.
 > 
 > Even rally_full is executed

Hi Cedric,

I added you as a reviewer to start a discussion about how to run Rally cases from Functest.

I would like to provide you with a description of what the blacklists are trying to achieve here.

What we would like to do is to have Functest Rally Full split into Testcases. The current form of Functest uses this term into a hardcoded list of 9 Testcases ['authentication', 'quotas', 'keystone', ...]. It also runs the sub-testcases of each Testcase but it does not give the user the option to run all sub-testcases of a specific Testcase. Furthermore, Functest uses a blacklist mechanism to exclude a specific quota sub-testcase and some functionality from nova and neutron sub-testcases.

Solution
This patch leverages the blacklist mechanism to separate the run of testcases of rally_full.
An example: In order to run only sub-testcases of Authenticate the user should blacklist all the sub-testcases of all 8 other Testcases. If a user runs from dovetail functest.rally.authenticate then he is going to run only Authenticate* sub-testcases.

Furthermore, the original blacklist of Functest is respected, because if you have a look to quotas, neutron and nova black_lists you will see the existing functest rules. So Dovetail is conformed to them, it respects them and it does not change them.

Outcome
A user can be verified that it passes parts of functest rally_full but in order to pass all functest rally_full then he must have PASS to all:

functest.rally.authenticate
functest.rally.authenticate
functest.rally.authenticate

Patch-set: 4