Update patch set 10
Patch Set 10:

Rally_full is always successful and you should check first the functest gates to avoid mixing installer and functest issues.
https://build.opnfv.org/ci/job/functest-opnfv-functest-benchmarking-latest-rally_full-run/

It should be noted that we don't run migration tests in our pod simply because we only have an all-in-one one from the time being.

Rally full is designed to provide a full confidence (it ends in less than 2 hours). You may select rally_sanity instead but the target is different.

Patch-set: 10