Update patch set 10
Patch Set 10:

(1 comment)

> And I don't understand why its duration could be an issue here
 > (it's a certification not a patch verification).
 > 
 > And there are lots of ways to decrease the overall duration.
 > Our full gate (docker builds + rally_full included) ends after 5
 > hours which is much lesser that the OPNFV installers:
 > https://build.opnfv.org/ci/view/functest/job/functest-latest-gate/79/

I agree with Cedric here. The reason I created the patch with rally_full instead of rally_sanity is that we want to certify that a platform can perform well no matter how much time do we need to spend on tests. Yes indeed, rally_full needs a lot of time comparing to rally_sanity but I believe this is the correct one to use.

Patch-set: 10