Update patch set 10
Patch Set 10:

(1 comment)

I totally agree that we shouldn't choose simple test cases in order to reduce the time cost. Reducing duration time is not the prior target of OVP. However, we need to know why it takes so long and if it is possible to save some time.
I found that it may be the iteration=10 which makes it take more than 2 hours. For example (maybe not correct), if iteration is 1, then it takes 30 min. So 10 iterations take 5 hours.
I am curious about the reason to set it run 10 iterations. For functional tests, it is ok to run every test only one time to ensure it works. What's the point of running it 10 times?

Patch-set: 10
Label: Verified=0