Update patch set 5
Patch Set 5:

I tested it to the remote server and it works.

However, it needed to use /vvp directory for operations because the vvp team changed the container to use vvpuser instead of root. As a result I was receiving permission denied errors.

This is something I found out and was not informed about it.

In addition, the current vvp docker image misses an apk package (tk) which makes it unable to run the validation scripts.

Patch-set: 5
Label: Verified=0