Fix the verdict of TC in sub-testcase list result
When the sub-test case list is defined, in the dovetail results,
each test case is appeared with the its verdict next to its name.
Once a sub-testcase is passed, the verdict PASS is appeared next
to the name of sub testcase.
Once a sub-testcase is skipped or failed, the verdict FAIL is
appeared next to the name of sub-testcase.

This fix updates the code so that the verdict of skipped and failed
sub test cases to be appeared respectively next to its sub-tsetcase
name.

Change-Id: Ib05b3793a57acbaa67b89321a840ae11251c2417
Signed-off-by: Panagiotis Karalis <pkaralis@intracom-telecom.com>