Update patch set 2
Patch Set 2:

> (1 comment)

I agree with that, but I left this check to run also for 'Success' cases as improvement.
If all sub testcases of a class is passed and this method is run on class level and not for each sub testcase, the method should return true for the entire class.

Patch-set: 2
Reviewer: Gerrit User 7167 <7167@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0