Update patch set 2
Patch Set 2: -Code-Review

I switch to 0 because it cleans useless deps and it helps as a first step.

Then I would suggest to automatically sync the requirements as I'm doing for few projects. But it has to be completed in a second change to add the upper-constraints in the pip install calls.

Patch-set: 2
Reviewer: Gerrit User 3405 <3405@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=0