Update patch set 3
Patch Set 3: Code-Review-1

I have tried to create a 'results' directory with non root user but it still got the permission denied error.
The best solution I can find is that, the 'results' directory can be created by any user (root or non root), but need to change the mode of it to be 777. It works for me with both root and non root. So we need to add this to the user guide.
Another problem is in report OnapVvpCrawler, need to only crawl from file_path[0].

Patch-set: 3
Label: Code-Review=-1