Update patch set 1
Patch Set 1:

> (2 comments)

The form (docusign legal), that are you mentioned, is not the application form that we have added in the portal and the user is responsible to fulfilled as part of a new application for new results. right?
As part of design of  the ONAP portal, we have prepared an application form. First, the user has to upload results and for each result the user can submit an application form requesting a certification / badge.
Moreover, this was a decision also for the general workflow.
Am I correct?

Furthermore, during the application form completion, the user could upload also a logo.

My understanding for those forms is: the 'form (docusign legal)' is different from application form.
- The reason of the 'form (docusign legal)' is to grant someone access to upload results and submit applications as a user. 
- And the reason of application form is to submit more info along with relevant results during to upload of new results from a user (this user has already the proper permission, taking it from previous step), requesting a certification /badge
Am I right?

If you agree with me, maybe the step 3 should be updated and add that the user should submit an application form along with each result in order for a proper review to be done (requesting the badge for the relevant results).
Reminder: According to the old workflow this application is fulfilled by administrator, but now by user.

Concluding, the same workflow is applied also on OVP portal and I have pushed a patch about this and I have added you as reviewer [1].

[1] https://gerrit.opnfv.org/gerrit/#/c/66617/

Patch-set: 1