Update patch set 1
Patch Set 1:

With the change to the VNF portal, we combined the application and the results upload.  We could change it so anyone can start the process, but we won't be separating the application from the results upload part.  To do this, we need LF to give the OK we can store those results without the participation form and we would need to change the permission scheme of the portal a little bit, to allow anyone to do that application, but only the "user" (i.e. someone with a completed participation form) to submit the application for review.  Is this what we want to do / ask for as a process?

Patch-set: 1