Update patch set 9
Patch Set 9:

> Why don't you consider running SFC tests as proposed by OpenStack
 > and then Functest? They can be covered by Functest gates as it was
 > few weeks ago.
 > 
 > Does the ODL part still maintain?

To be frank, I'm not sure whether the ODL part is still maintain and I'd like to investigate further this one.
Moreover, I discussed with Dan whether this TC will give a real benefit on the Dovetail project. Therefore, Dan agreed that we should 'pause' a little bit the integration of SFC test cases (at least, until be stable the strategy of the next release). Keeping also our eyes open for the participation of SFC project on the next release.

Patch-set: 9