Fault fixing for cmd_exec util method
When the 'dovetail run' command is assigned, the error
"TypeError: must be str, not bytes" is appeared.

Due to a different handling of strings in the Python3, this error
is appeared during of command execution (i.e. exec_cmd method).

Each piece of the string should be decoded, before it will be
appended to the rest one.

Signed-off-by: Panagiotis Karalis <panos.pkaralis@gmail.com>
Change-Id: I65629f3f76cc3e44f3926a744d00791ef588d2aa