Update patch set 2
Patch Set 2: Code-Review-1

(15 comments)

Thanks for pulling this together.  I've made a number of comments throughout. My general concern is that in spots (i've noted them) it is not clear there are different badges, and both what is being tested and how varies.  We may need some general prefaces that provides that overview so as a reader you know how to navigate the material.

I've also posted some comments about what is documented in ONAP vs. the OVP portal.  Given that releases of OPNFV and ONAP are not synchronized, I think the material related to the test suites contributed in ONAP would be better served to be documented in ONAP and linked to.

Happy to discuss further on the next CVC call.

Patch-set: 2
Label: Code-Review=-1